+++ src/Ruleset/RuleCountry.cpp
@@ -118,7 +118,7 @@ double RuleCountry::getLabelLatitude() const
  */
 bool RuleCountry::insideCountry(double lon, double lat) const
 {
-	for (unsigned int i = 0; i < _lonMin.size(); ++i)
+	for (size_t i = 0; i < _lonMin.size(); ++i)
 	{
 		bool inLon, inLat;
